Rename 'ai' -> 'enemies'
