+++ src/game_logic/enemies/blue_guard.cpp
