+++ src/game_logic/enemies/bomber_plane.cpp
