+++ src/game_logic/enemies/ceiling_sucker.hpp
