+++ src/game_logic/enemies/floating_laser_bot.cpp
