+++ src/game_logic/enemies/red_bird.hpp
