+++ src/game_logic/enemies/rigelatin_soldier.cpp
