+++ src/game_logic/enemies/security_camera.cpp
