+++ src/game_logic/enemies/simple_walker.cpp
