+++ src/game_logic/enemies/smash_hammer.hpp
