+++ src/game_logic/enemies/spike_ball.hpp
