+++ src/game_logic/entity_factory.cpp
@@ -22,34 +22,34 @@
 #include "engine/physics_system.hpp"
 #include "engine/sprite_tools.hpp"
 #include "game_logic/actor_tag.hpp"
-#include "game_logic/ai/blue_guard.hpp"
-#include "game_logic/ai/bomber_plane.hpp"
-#include "game_logic/ai/boss_episode_1.hpp"
-#include "game_logic/ai/ceiling_sucker.hpp"
-#include "game_logic/ai/eyeball_thrower.hpp"
-#include "game_logic/ai/floating_laser_bot.hpp"
-#include "game_logic/ai/hover_bot.hpp"
-#include "game_logic/ai/laser_turret.hpp"
-#include "game_logic/ai/messenger_drone.hpp"
-#include "game_logic/ai/prisoner.hpp"
-#include "game_logic/ai/red_bird.hpp"
-#include "game_logic/ai/rigelatin_soldier.hpp"
-#include "game_logic/ai/rocket_turret.hpp"
-#include "game_logic/ai/security_camera.hpp"
-#include "game_logic/ai/simple_walker.hpp"
-#include "game_logic/ai/slime_blob.hpp"
-#include "game_logic/ai/slime_pipe.hpp"
-#include "game_logic/ai/smash_hammer.hpp"
-#include "game_logic/ai/snake.hpp"
-#include "game_logic/ai/spider.hpp"
-#include "game_logic/ai/spike_ball.hpp"
-#include "game_logic/ai/watch_bot.hpp"
 #include "game_logic/behavior_controller.hpp"
 #include "game_logic/collectable_components.hpp"
 #include "game_logic/damage_components.hpp"
 #include "game_logic/dynamic_geometry_components.hpp"
 #include "game_logic/effect_actor_components.hpp"
 #include "game_logic/effect_components.hpp"
+#include "game_logic/enemies/blue_guard.hpp"
+#include "game_logic/enemies/bomber_plane.hpp"
+#include "game_logic/enemies/boss_episode_1.hpp"
+#include "game_logic/enemies/ceiling_sucker.hpp"
+#include "game_logic/enemies/eyeball_thrower.hpp"
+#include "game_logic/enemies/floating_laser_bot.hpp"
+#include "game_logic/enemies/hover_bot.hpp"
+#include "game_logic/enemies/laser_turret.hpp"
+#include "game_logic/enemies/messenger_drone.hpp"
+#include "game_logic/enemies/prisoner.hpp"
+#include "game_logic/enemies/red_bird.hpp"
+#include "game_logic/enemies/rigelatin_soldier.hpp"
+#include "game_logic/enemies/rocket_turret.hpp"
+#include "game_logic/enemies/security_camera.hpp"
+#include "game_logic/enemies/simple_walker.hpp"
+#include "game_logic/enemies/slime_blob.hpp"
+#include "game_logic/enemies/slime_pipe.hpp"
+#include "game_logic/enemies/smash_hammer.hpp"
+#include "game_logic/enemies/snake.hpp"
+#include "game_logic/enemies/spider.hpp"
+#include "game_logic/enemies/spike_ball.hpp"
+#include "game_logic/enemies/watch_bot.hpp"
 #include "game_logic/interactive/blowing_fan.hpp"
 #include "game_logic/interactive/elevator.hpp"
 #include "game_logic/interactive/enemy_radar.hpp"
