+++ src/game_logic/ingame_systems.hpp
@@ -23,22 +23,22 @@
 #include "engine/physics_system.hpp"
 #include "engine/rendering_system.hpp"
 #include "engine/rendering_system.hpp"
-#include "game_logic/ai/blue_guard.hpp"
-#include "game_logic/ai/hover_bot.hpp"
-#include "game_logic/ai/laser_turret.hpp"
-#include "game_logic/ai/messenger_drone.hpp"
-#include "game_logic/ai/prisoner.hpp"
-#include "game_logic/ai/rocket_turret.hpp"
-#include "game_logic/ai/simple_walker.hpp"
-#include "game_logic/ai/slime_blob.hpp"
-#include "game_logic/ai/spider.hpp"
-#include "game_logic/ai/spike_ball.hpp"
 #include "game_logic/behavior_controller_system.hpp"
 #include "game_logic/camera.hpp"
 #include "game_logic/damage_infliction_system.hpp"
 #include "game_logic/debugging_system.hpp"
 #include "game_logic/dynamic_geometry_system.hpp"
 #include "game_logic/effects_system.hpp"
+#include "game_logic/enemies/blue_guard.hpp"
+#include "game_logic/enemies/hover_bot.hpp"
+#include "game_logic/enemies/laser_turret.hpp"
+#include "game_logic/enemies/messenger_drone.hpp"
+#include "game_logic/enemies/prisoner.hpp"
+#include "game_logic/enemies/rocket_turret.hpp"
+#include "game_logic/enemies/simple_walker.hpp"
+#include "game_logic/enemies/slime_blob.hpp"
+#include "game_logic/enemies/spider.hpp"
+#include "game_logic/enemies/spike_ball.hpp"
 #include "game_logic/interactive/elevator.hpp"
 #include "game_logic/interactive/enemy_radar.hpp"
 #include "game_logic/interactive/sliding_door.hpp"
